requirements/base.txt
# Django
django==1.11.3

# Configuration
dj-database-url==0.4.1
python-decouple==3.0
whitenoise==3.2.2

# Django 3rd party modules
django-allauth==0.30.0
django-extensions==1.7.5
django-model-utils==2.6
django-rest-auth==0.9.0
django-webpack-loader==0.4.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2
django-crispy-forms

# Time zones support
pytz==2016.10

pillow

stripe
